Add AGENTS.md file and associated changes.
